+++ bin/parser.rb
@@ -1,6 +1,2 @@
 # ruby parser for 835 file
-require_relative '../lib/segment'
-require_relative '../lib/isa'
-require_relative '../lib/file_835'
-
-# Look into using Module.const_get(str)
+require_relative '../lib/X221/segment'
